Refactored training code into a separate function
